move int tests to another file
